requirements.txt
opencv_python==4.5.4.60
Pillow==8.4.0
PySimpleGUI==4.55.1
requests==2.26.0
tqdm==4.62.3
